@@ -1,51 +1,45 @@
 {
-  "name": "react-frontend-dev-portfolio",
-  "homepage": "https://dorota1997.github.io/react-frontend-dev-portfolio/",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "bootstrap": "^4.5.2",
-    "jquery": "^3.5.1",
-    "react": "^16.13.1",
-    "react-awesome-slider": "^4.1.0",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-ga": "^3.1.2",
-    "react-scripts": "3.4.3",
-    "react-switch": "^5.0.1",
-    "react-typical": "^0.1.3",
-    "react-vertical-timeline-component": "^3.3.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@iconify/icons-logos": "^1.0.12",
-    "@iconify/react": "^1.1.3",
-    "gh-pages": "^3.1.0",
-    "sass": "^1.42.1"
-  }
+    "name": "react-portfolio",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "bootstrap": "^5.2.3",
+        "emailjs-com": "^3.2.0",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.7.0",
+        "react-dom": "^18.2.0",
+        "react-helmet-async": "^1.0.7",
+        "react-icons": "^4.1.0",
+        "react-router-dom": "^6.6.2",
+        "react-scripts": "^5.0.1",
+        "react-transition-group": "^4.4.2",
+        "typewriter-effect": "^2.18.2",
+        "web-vitals": "^3.1.1"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject",
+        "predeploy": "yarn build && cp build/index.html build/404.html",
+        "deploy": "gh-pages -d build"
+    },
+    "eslintConfig": {
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    }
 }
